requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
attrs==21.4.0
autoflake8==0.3.2
autopep8==1.6.0
awscli==1.25.3
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
blackcellmagic==0.0.3
bleach==5.0.0
botocore==1.27.3
catboost==1.0.6
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-option-group==0.5.3
colorama==0.4.4
colour==0.1.5
cookiecutter==2.1.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
e==1.4.5
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
flake8==4.0.1
fonttools==4.33.3
graphviz==0.20
greenlet==1.1.2
idna==3.3
ipykernel==6.13.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.0
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.0
jupyter==1.0.0
jupyter-black==0.3.1
jupyter-client==7.3.3
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lab==7.0
logger==1.4
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbautoexport==0.5.0
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
ohio==0.5.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydot==1.4.2
pydotplus==2.0.2
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests==2.27.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.1.1
scikit-learn-extra==0.2.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.2.0
styleframe==4.1
tenacity==8.0.1
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenize-rt==4.2.1
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
tweet-preprocessor==0.6.0
txt2tags==3.7
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
wordcloud==1.8.1
xgboost==1.6.1
xlrd==2.0.1